requirements.txt
Flask>=2.2
requests>=2.30
python-dotenv>=1.1
